Sync upstream (#3)
* First try at HTTPS_METHOD

* Implemented HTTPS noredir

* Documented HTTPS_METHOD

* Semantic change to trigger CI rebuild

* Trying to trigger CI

* UPGRADE docker-gen release from 0.4.2 to 0.5.0

* Use new Network interface to support new overlay network

* Update README - Nginx proxy must use host network to support overlay network

* Complete documentation

* Update README wording for overlay networking

* changed error code for non-usable/default SSL cert, fixes #341

* Support container in one network shared with current container

* [TEMPLATE] fix variable call

* Upgrade docker-gen from 0.5.0 to 0.7.0

* Test to support overlay networks

* Unit test - removed network creation

* moved nginx to 1.9.12

* Update to docker-gen 0.7.0

Since [1] some timings seem to have changed. This caused the unit tests
to fail intermittently, from my testings especially on Ubuntu systems
(much less often on e.g. Arch).

This commit adds the `dockergen_wait_for_event` helper-function to try
and wait for the configuration to be generated by docker-gen before
continuing on with the actual tests themselves.

Additionally, at the end of every test file, all containers spun up by
the bats-tests will be stopped. This required adding the `bats-type`
label to every container started during the bats-tests.

The stopping of the containers reduces the amount of events docker-gen
has to process, thus resulting in lower wait times for the generation to
happen.

[1]: https://github.com/jwilder/docker-gen/commit/50435652b16b3874a5907765d33fa31150876d92

* [FIX] expose port for separated containers test

* Upgrade Docker dependency from 1.9 to 1.10

* Switch forego downloads

Official ddollar site has been down for a while and not sure if it's
coming back up or not.

* Add travis.yml

From #312

* Added SSL/HTTPS tests

* Normalized indentation, added explicit CERT_NAME

* Using different path method to hopefully fix cert volume issue in CI systems

* Add docker hub badges

* Update to docker-gen 0.7.1

* Remove circle.yml

Using travis now

* Fix a typo in comment

* Expanded documentation on HTTPS_METHOD

* Removed HSTS when HTTPS_METHOD=noredirect, added tests, improved docs wrt HSTS

* Update nginx 1.9.15

* Remove docker compose v2 notes

* Remove Circle CI badge

* Fix template error when /etc/nginx/certs does not exist

* Remove -only-exposed from Procfile

* Add docker-compose.yml v2 example

* Remove -only-exposed from separate container instructions

* Revert 9c93efa

* Update to docker-gen 0.7.3